Replaced Webdriver stuff in register with method. Moved WebDriverFactory.cs
